fix: rename main presenter
